Disallow passing functions that don't return Tensors to vmap
I overlooked this in the initial vmap frontend api PR. Right now we
want to restrict vmap to taking in functions that only return Tensors.
A function that only return tensors can look like one of the following:
```
def fn1(x):
    ...
    return y

def fn2(x):
    ...
    return y, z
```
fn1 returns a Tensor, while fn2 returns a tuple of Tensors. So we add a
check that the output of the function passed to vmap returns either a
single tensor or a tuple of tensors.

NB: These checks allow passing a function that returns a tuple with a
single-element tensor from vmap. That seems OK to me.

Test Plan:
- `python test/test_vmap.py -v`

ghstack-source-id: ec408e24fecc70e30b698c1f827392ffaf0b03bb
Pull Request resolved: https://github.com/pytorch/pytorch/pull/40518